Move DynamoDB helper functions to a separate file
